% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confeti_functions.R
\name{confeti}
\alias{confeti}
\title{Main function to run CONFETI.}
\usage{
confeti(expr_data, snp_data, return_all = TRUE, cores = 1, k = NULL,
  var_cutoff = 95, ica_runs = 1, scale_pheno = FALSE,
  h_clust_cutoff = 0.3, similarity_measure = "peaks", threshold = 0.05,
  return_pval = FALSE)
}
\arguments{
\item{expr_data}{Expression matrix with dimensions g x n.}

\item{snp_data}{Genotype matrix with dimensions s x n.}

\item{return_all}{If TRUE, the full ICA results are returned. If set FALSE
only the sample covariance matrix and lower dimensional phenotype matrix are returned.}

\item{cores}{Number of cores to use for genotype association testing, and ICA multi runs (if ica_runs \> 1).}

\item{k}{Number of components to be estimated or method to estimate it.}

\item{var_cutoff}{Percent variance threshold to use when <k_est> is not supplied.}

\item{ica_runs}{Number of times to run ICA. If this is set to a number larger than 1,
only ICs that replicate between runs are going to be returned}

\item{scale_pheno}{If set to TRUE the pre-processing step of the data will include a scaling step.
This will divide all phenotypes by their standard deviation.}

\item{h_clust_cutoff}{is the cutoff value used in hierarchical clustering. Default is set to 0.3.}

\item{similarity_measure}{How to measure the similarity between ICs.}

\item{threshold}{Bonferroni significance threshold for genotype association testing}

\item{return_pval}{If set to TRUE, p-values for genotype ICA association testing are returned}
}
\value{
N x N covariance structure matrix.
}
\description{
Creating a sample covariance matrix based on non-genetic independent components.
}
\keyword{keywords}

