% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confeti_functions.R
\name{get_similarity_mx}
\alias{get_similarity_mx}
\title{Function to calculate the covariance structure (K) matrix.}
\usage{
get_similarity_mx(ica_object, normalize_ystar = FALSE)
}
\arguments{
\item{ica_object}{ICA object that contains the confounding factor labels generated with `ica_genotype_test`.}

\item{normalize_ystar}{If set to TRUE phenotype values of the reconstructed y will be mean centered and divided by their standard deviation.}
}
\value{
N x N covariance structure matrix.
}
\description{
Creating the K matrix based on non-genetic independent components.
}
\keyword{keywords}

