% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confeti_functions.R
\name{ica_genotype_test}
\alias{ica_genotype_test}
\title{Function that uses lrgpr's glmApply to test associations between IC coefficients and genotypes.}
\usage{
ica_genotype_test(ica_object, genotype_mx, n_cores = 1,
  sig_threshold = 0.05, return_pval = FALSE)
}
\arguments{
\item{ica_object}{ICA object that is created with the `run_ica()` function from the `picaplot` package.}

\item{genotype_mx}{Input genotype matrix with N samples x G genotypes.}

\item{n_cores}{Number of cores to use for this process.}

\item{sig_threshold}{Bonferroni significance threshold to use in calling significant p-values.}

\item{return_pval}{TRUE or FALSE value indicating whether the calculated p-value matrix should be returned.}
}
\value{
ICA object with confounding factor information added.
}
\description{
Testing associations between IC coefficients and genotypes.
}
\keyword{keywords}

